requirements.txt
pyserial
numpy
pandas
scipy